Move relevant functions to scheduler.c
